timeseries fix
