+++ src/autogluon/cloud/predictor/multimodal_cloud_predictor.py
@@ -1,5 +1,7 @@
 import logging
+from typing import Dict
 
+from ..backend.constant import MULTIMODL_SAGEMAKER, SAGEMAKER
 from .cloud_predictor import CloudPredictor
 
 logger = logging.getLogger(__name__)
@@ -14,6 +16,13 @@ class MultiModalCloudPredictor(CloudPredictor):
         Type of the underneath AutoGluon Predictor
         """
         return "multimodal"
+
+    @property
+    def backend_map(self) -> Dict:
+        """
+        Map between general backend to module specific backend
+        """
+        return {SAGEMAKER: MULTIMODL_SAGEMAKER}
 
     def _get_local_predictor_cls(self):
         from autogluon.multimodal import MultiModalPredictor
